Moved starFilterHelper back to App.jsx
